requirements.txt
pbr>=5.4.3 # Apache-2.0
six>=1.11.0
Paste>=2.0.3
WebOb>=1.7.4
PasteDeploy>=1.5.2
keystoneauth1>=3.4.0 # Apache-2.0
keystonemiddleware>=4.21.0
oslo.middleware>=3.35.0
oslo.log>=3.36.0
oslo.config>=5.2.0
oslo.service>=1.30.0 # Apache-2.0
oslo.i18n>=3.15.3 # Apache-2.0
python-memcached>=1.59
python-neutronclient>=6.7.0
